Filename prefixed image directory name
An option to put all the images of current markdown file in a
directory with current filename without the extension prefixed to
`mdip_imgdir`with an underscore in between.
This is supposed to help organize images related to multiple
markdown documents in the same directory.

For more customization, use `expand()` inside `mdip_imgdir`
Eg:
```vim
" The following will save all the images in the directory
../all_images/currentfilename_without_extension
let g:mdip_imgdir = '../all_images' . expand('%:r')
```